requirements.txt
# Core dependencies
tqdm
python-dotenv
openai
numpy
loguru
PyMuPDF

# 需要高硬件资源
mineru[all]>=2.1.9
xinference>=1.8.0